git-merge-pr: Reformat PR message markdown for commit
The markdown in the PR body is not completely suitable for a commit
message. Notably are headings starting with a hash (#), which appear as
comments in the git commit message. Convert those headings to underline
style.

Also remove triple-backticks for code blocks and convert them to
indented blocks. These render the same way with markdown but are nicer
for a commit message.