git-merge-pr: Fix one case of merging from fork
When merging from a fork, the `-h` argument to `hub pr show` needs to be
qualified with the owner repository. I don't yet know how to find that
out, but that is only needed if you are merging to your current branch,
not merging from your current branch. This latter form is the most used,
so get rid of the `-h <branch>` arg altogether in that case. This makes
it work again.

It is still broken if you are using the first form and are working with
the fork model. I need to figure out how to find the "owner" of the
fork.