git-merge-pr: Fix count/list of commits on branch
Using `...` to see the commits on the branch to exclude the ones on the
other branch does not work. It does work with `git diff` but not `git
log`.

e.g. if you have the following commit graph:

    master --> C   Z <--- branch
               |   |
               B   Y
                \ /
                 A

If you run `git diff master...branch`, it excludes commits `B` and `C`
from the diff. If you run `git log master...branch` it includes commits
`B` and `C`. For `git log`, you need to use `..`: `git log
master..branch`. I no understand.

So fix `git-merge-pr` to use `..` as it uses it with `git log`.

This normally would not be noticed as it is only an issue if master has
advanced past your branch and a pre-check stops you merging in that
case. However, I want to make that a flag because sometimes I like to
merge non-up-to-date branches because I know it doesn't matter.