git-merge-pr: Include PR body in single commit PRs
The git-merge-pr script left out the PR message from the merge commit
when there was a single commit in the PR as in that case the PR message
is usually just the single commit message. But it still makes sense to
duplicate that in the merge commit so that the merge commit can still
describe what is going on. So always include the PR message in the merge
commit.