git-merge-pr: Fix diff stat in merge message
Commit fba3f715 (Fix count/list of commits on branch) changed git log,
diff and rev-list to use `..` instead of `...`. As described in the
commit message of that commit, that change should not have been made for
`git diff`. Revert that part of the commit so `git diff --stat` includes
a correct stat.

Thanks to Julia for spotting this.

The difference between `..` and `...` for `diff`/`log` is described at
https://stackoverflow.com/questions/53120060/why-double-dot-vs-triple-dot-in-git-behaves-different-for-diff-and-log