git-pr-merge: Rename git-merge-pr to git-pr-merge
Rename git-merge-pr to git-pr-merge. This is in anticipation of adding
other "pr" subcommands, such as `git pr update`, etc and implementing
that as a `git-pr` binary.

Change function names in the shell script to reflect the new name.

Migrate the existing config section "merge-pr" to "pr.merge". This
migration will remain for a little while until all users update and
migrate all their repos.