git-pr-merge: Add symlink from git-merge-pr
Add a symlink from git-merge-pr to git-pr-merge so muscle memory can
slowly change from the rename in the previous commit. This was not done
in the previous commit as git would not properly detect the rename as
this was seen as a typechange and an add of the renamed script.