git-credential-subpath: Add credential helper
Add a credential helper to allow usernames to be specified on a repo
subpath basis. gitcredentials(7) only work on the hostname or the full
repo path, not a subpath, such as https://github.com/org. If you want to
use different usernames for different github orgs, you can't with stock
gitcredentials.

This allows you to add a config sections such as:

    [credential]
        username = octocat1
    [credential "https://github.com/org"]
        username = octocat2

where the username `octocat1` will be used by default, but `octocat2`
will be used for any repositories under `https://github.com/org`

This helper will chain to another helper that will actually store the
passwords.