git-pr-merge: Set variables to default, dont use git-config
Set the shell variables to their defaults rather than letting
`git-config` do it with the `--default` parameter. This makes the
variables properly initialised to valid values so if the script is
sourced and config values not read from git-config, the variables have
valid values. It just seems a bit cleaner too and simplifies the code.