git-pr-merge: Use PR title as merge commit title
Use the PR title as the merge commit title instead of saying "merge
branch-name", because branch names are usually not all that descriptive
and do not usually make good English sentences.

Add a line before the commit list that indicates it is a merge commit
since we've lost the "Merge" at the start of the commit.

Put the pull request URL in a trailer in the commit message. Still to do
is identify trailers in the PR message and move them to the end of the
commit.

Rework the script a little so the internal functions are more
self-contained and do not depend on the environment being set up. This
makes it easier to source the script at a shell to test those internal
functions.