git-pr-merge: Merge commit from repo root
Change directory to the repo root before merging the branch. This is
necessary as the current directory may not exist on master when the
currently-being-merged branch has added it. When we switch back to
master to merge the branch, we end up in a "detached" directory. How
this is handled can depend on whether it is a local or network
filesystem, or depending on the OS.

By changing to the repo root directory, we avoid any of this
uncertainty. We change back once the merge is complete, as that
directory will exist then.