git-pr-merge: Add some help text to configure script
When a remote branch deletion fails during clean up, this is usually
because GitHub is set up to automatically delete the branch on merge. If
it does fail, output an informative message with a `git config` command
to turn off branch deletion on the repo that can be easily cut and
paste.

Also fix up the header comment while we're here.