git-pr-merge: Simplify var names, remove parameter
Simplify the variable names:
* merge_from_branch -> feature
* merge_to_branch -> master

Technically the merge-to branch is not necessarily master, but most
often is (for foxygoat), so just call it master for better readability.
"base" might be a technically better term, but it is less known.

Remove the optional parameter to the script. I never ever use it, and it
just complicates things.