pr-merge: Add squash merging using github api
Make git-pr-merge do a squash merge via the github API if the git config
var pr.merge.squash is set to true. The script does not check if squash
merging is allowed on the repo, so it may fail.

Enable squash merging by running:

    git config --local pr.merge.squash true

Currently it is not possible to squash merge with a command line flag
(which would allow you choose on a per-pr basis).

The merge message for a squash merge commit omits the diff stat and the
list of commit titles, as the former is easily shown with `git log
--stat` and the latter makes little sense when all the commits are being
squashed as we lose the individual commits and there will either be only
1 or fixups that are not relevant.