git-pr-merge: Prune remote if not deleting remote branch
Prune branches from the remote we've pushed to if we are configured not
to delete the branch on the remote. Usually you do not delete the branch
on the remote as it is configured to be deleted on pr merge. But that
will leave the remote branch lying around in the local repo until you
next fetch. So run a "git remote prune <remote>" to clean up the remote
branch locally.