git-pr-merge: Add option for squash merging single commit PRs
Add a git config option - `pr.merge.squash-singles` - to perform a
squash merge when a PR has just one commit and to create a merge commit
when there are multiple commits on the PR.

This helps create a slightly cleaner history as single commits with a
merge commit seem to be more noisy than just a single commit on master.
Multiple commits are still grouped together by the merge commit which is
a useful thing to have.