🛀 git-pr-merge: Remove git-merge-pr compatibility (#27)
Remove the compatibility for git-merge-pr. This was the original name of
the script and was renamed to git-pr-merge. Some compatibility code as
well as a symlink were retained for the old name to keep working. This
has not been used for quite some time now, so compatibility is no longer
required.

Pull-Request: https://github.com/foxygoat/git-scripts/pull/27