git-pr-merge: Switch to `gh` from `hub` (#28)
🐙 git-pr-merge: Switch to `gh` from `hub` (#28)

Use GitHub's cli tool `gh` instead of the older `hub` tool. The latter
is not really developed any more, and the former is "official". It also
has more features.

Most of the conversion is just a mechanical translation, except the
squash merge which no longer needs to use the api as `gh` has a
`gh pr merge --squash` command, so we use that instead.

Pull-Request: https://github.com/foxygoat/git-scripts/pull/28

# Gitmoji: https://gitmoji.carloscuesta.me
# rxaviers list: https://gist.github.com/rxaviers/7360908
#