git-pr-merge: Squash merge using API, not `gh pr merge`
Use `gh api` to perform a squash merge instead of using `gh pr merge`.
The latter does not allow you to set the title of the squash merge
commit - it uses the PR title, but we want to add a title prefix.
Currently it "doubles up" the title.

This partially reverts 759e1b54402519aa190f68ef689a5af4132901e0 but
converts `hub api` to `gh api`, putting flags last and removing the
leading slash of the API, as this is what `gh api --help` says.