✨ git-pr-merge: Allow merging via PR number with -p (#31)
Add the `-p NN` flag to merge a PR given a PR number rather than merging
the currently checked-out branch. This is to support merging PRs raised
by others for which you do not have the branch, which is often on a fork
where the branch does not even exist in origin.

Pull-Request: https://github.com/foxygoat/git-scripts/pull/31