git-pr-merge: Strip reviewer comments from PR message
When constructing a merge commit message from a pull request body
(description), remove anything after a blank line followed by a line
solely of `---`. including the marker line. The `---` indicates the
start of reviewer comments of a PR and should not be included in the
merge commit message. This is the same as `git-am` does in a way - the
commit message comes before the `---` and the patch after. Patching
ignores any text before the formal start of the patch.