Add .git-blame-ignore-revs to ignore reformat commits
Add a `.git-blame-ignore-revs` file that can be used with `git blame` to
ignore commits recorded in that file - typically commits that just
format files.

Add the commit hash of the previous commit that reformatted the scripts
with `shfmt`.

This file will be used automatically by GitHub. It can be used with
`git blame` with the `--ignore-revs-file` flag or specify it in the
local repo config with

    git config --local blame.ignoreRevsFile .git-blame-ignore-revs

You cannot do that globally unfortunately as if a repo or a commit (say
if bisecting) does not have that file, `git blame` will error out saying
it cannot find the file.