✨ git-pr-merge: Check if GitHub will delete the branch (#35)
Check if the GitHub repo setting `deleteBranchOnMerge` is set before
attempting to delete the feature branch after merging it. Prior to this,
we would attempt to delete the branch, fail, and emit a warning message
with a suggestion to turn off auto branch deletion.

Now, you can set branch deletion globally so the feature branch gets
deleted when merged, but if `deleteBranchOnMerge` is on, we just no
longer attempt to delete the remote branch and do not emit the warning.

While we're here, fix up one remaining use of short flags with
`git branch` - just force-delete always so we no longer need to chose
whether to force or not. Makes the code simpler.

Pull-Request: https://github.com/foxygoat/git-scripts/pull/35