🚛 git-pr-merge: Move PR body trailers to end of merge commit (#36)
Move any trailers at the end of a PRs message body to the end of the
merge commit message so that the PR trailers appear in the correct spot
in the merge commit message, after the merge summary (log, diffstat).

We still add a `Pull-request:` trailer (note that the capital "R" in
request has been changed to lower case to conform to trailer
conventions), and that is bundled with any PR trailers.

Pull-request: https://github.com/foxygoat/git-scripts/pull/36